refactor: Convert wake-db to TypeScript with retry polling logic
Address PR feedback from gemini-code-assist and copilot:

- Convert script from CommonJS JavaScript to TypeScript
- Replace fixed 2s timeout with retry polling mechanism (max 40s)
- Add finally block to ensure prisma.() runs on all paths
- Fix formatting: add newline between closing brace and function call
- Update package.json scripts to use 'tsx' instead of 'node'
- Add proper TypeScript error handling with type guards

The retry mechanism polls the database with exponential backoff instead of
assuming it will be ready in a fixed time, making the script more robust
for variable Neon wake-up times.
