+++ package.json
@@ -4,7 +4,7 @@
   "private": true,
   "scripts": {
     "dev": "next dev --turbopack",
-    "dev:wake": "node scripts/wake-db.js && bun run dev",
+    "dev:wake": "tsx scripts/wake-db.ts && bun run dev",
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
@@ -21,7 +21,7 @@
     "db:migrate:reset": "prisma migrate reset",
     "db:generate": "prisma generate",
     "db:seed": "tsx prisma/seed.ts",
-    "db:wake": "node scripts/wake-db.js",
+    "db:wake": "tsx scripts/wake-db.ts",
     "validate-env": "tsx scripts/validate-env.js"
   },
   "dependencies": {
